moving all participant data to separate ParticipantData model
