+++ lab/experiments/templates/experiments/invite.html
@@ -50,10 +50,10 @@
                     <thead>
                         <tr>
                             <th>
-                                {% get_verbose_field_name "participants" "Participant" "name" %}
+                                {% get_verbose_field_name "participants" "ParticipantData" "name" %}
                             </th>
                             <th>
-                                {% get_verbose_field_name "participants" "Participant" "email" %}
+                                {% get_verbose_field_name "participants" "ParticipantData" "email" %}
                             </th>
                             <th>
                                 {% trans 'experiments:invite:last_call' %}
