+++ lab/experiments/templates/experiments/participants_table.html
@@ -13,7 +13,7 @@
                 {% trans 'timeslots:day' %}
             </th>
             <th>
-                {% get_verbose_field_name "participants" "Participant" "name" %}
+                {% get_verbose_field_name "participants" "ParticipantData" "name" %}
             </th>
             <th>
                 {% trans 'experiments:participants:email' %}
