+++ lab/survey_admin/templates/survey_admin/invite.html
@@ -25,10 +25,10 @@
                         <tr>
                             <th></th>
                             <th>
-                                {% get_verbose_field_name "participants" "Participant" "name" %}
+                                {% get_verbose_field_name "participants" "ParticipantData" "name" %}
                             </th>
                             <th>
-                                {% get_verbose_field_name "participants" "Participant" "email" %}
+                                {% get_verbose_field_name "participants" "ParticipantData" "email" %}
                             </th>
                             <th>
                                 {% trans 'experiments:invite:last_call' %}
